move tests into version subdirectories
