+++ phpunit.xml
@@ -5,11 +5,13 @@
          colors="true"
 >
     <testsuites>
-        <testsuite name="Unit">
-            <directory>tests/Unit</directory>
+        <testsuite name="v1">
+            <directory>tests/v1/Feature</directory>
+            <directory>tests/v1/Unit</directory>
         </testsuite>
-        <testsuite name="Feature">
-            <directory>tests/Feature</directory>
+        <testsuite name="v2">
+            <directory>tests/v2/Feature</directory>
+            <directory>tests/v2/Unit</directory>
         </testsuite>
     </testsuites>
     <coverage>
