+++ tests/v1/Unit/BinaryAvailabilityTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Unit;
+namespace Tests\v1\Unit;
 
 use FFMpeg\FFMpeg;
 use Imagick;
