+++ tests/v1/Unit/CreateUserCommandTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Unit;
+namespace Tests\v1\Unit;
 
 use App\Console\Commands\CreateUser;
 use App\Models\User;
